requirements.txt
flask
flask-cors
pyaudio
wave
google-generativeai
groq
pandas
numpy
datetime
uuid
